Add Reviews in deteils of assembly

+++ PCAssembly/Controllers/UserAssemblyComponentsController.cs
@@ -1,42 +1,64 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using PCAssembly.Data;
+using PCAssembly.Services;
 
 namespace PCAssembly.Controllers
 {
     public class UserAssemblyComponentsController : UserController
     {
-        public UserAssemblyComponentsController(ComputerAssemblyContext context) : base(context) { }
-        public async Task<IActionResult> AssemblyComponents(int id)
+        private readonly AssemblyService _assemblyService;
+        private readonly AuthorizationService _authorizationService;
+
+        public UserAssemblyComponentsController(ComputerAssemblyContext context)
+            : base(context)
         {
-            var components = await _context.AssemblyComponents
-                                            .Where(ac => ac.AssemblyId == id)
-                                            .Include(ac => ac.Component)
-                                            .Select(ac => new ComponentViewModel
-                                            {
-                                                ComponentId = ac.Component.ComponentId,
-                                                Name = ac.Component.Name,
-                                                Description = ac.Component.Description,
-                                                Price = ac.Component.Price,
-                                                AssemblyComponentId = ac.AssemblyComponentId
-                                            })
-                                            .ToListAsync();
-
-            ViewBag.AssemblyId = id;
-
-            return View(components);
+            _assemblyService = new AssemblyService(context);
+            _authorizationService = new AuthorizationService();
         }
 
+        public async Task<IActionResult> AssemblyComponents(int id)
+        {
+            var assembly = await _assemblyService.GetAssemblyByIdAsync(id);
+
+            if (assembly == null)
+            {
+                return NotFound("Сборка не найдена.");
+            }
+
+            var currentUserId = User.FindFirst(System.Security.Claims.ClaimTypes.NameIdentifier)?.Value;
+            var isOwner = _authorizationService.IsOwner(assembly, currentUserId);
+
+            var components = await _assemblyService.GetComponentsAsync(id);
+            var reviews = await _assemblyService.GetReviewsAsync(id);
+
+            var model = Tuple.Create(assembly.AssemblyName, components, reviews, isOwner);
+            return View(model);
+        }
 
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> RemoveComponentFromAssembly(int assemblyComponentId)
         {
-            var assemblyComponent = await _context.AssemblyComponents.FindAsync(assemblyComponentId);
+            var currentUserId = User.FindFirst(System.Security.Claims.ClaimTypes.NameIdentifier)?.Value;
+
+            if (currentUserId == null)
+            {
+                return Unauthorized();
+            }
+
+            var assemblyComponent = await _context.AssemblyComponents
+                                                  .Include(ac => ac.Assembly)
+                                                  .FirstOrDefaultAsync(ac => ac.AssemblyComponentId == assemblyComponentId);
 
             if (assemblyComponent == null)
             {
                 return NotFound("The selected component connection was not found.");
+            }
+
+            if (!_authorizationService.IsOwner(assemblyComponent.Assembly, currentUserId))
+            {
+                return Forbid("You do not have permission to modify this assembly.");
             }
 
             _context.AssemblyComponents.Remove(assemblyComponent);
@@ -44,13 +66,5 @@ namespace PCAssembly.Controllers
 
             return RedirectToAction("AssemblyComponents", new { id = assemblyComponent.AssemblyId });
         }
-    }
-    public class ComponentViewModel
-    {
-        public int ComponentId { get; set; }
-        public string Name { get; set; }
-        public string Description { get; set; }
-        public decimal Price { get; set; }
-        public int AssemblyComponentId { get; set; }
     }
 }
